+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - fixed issue with null first and group by
 - fixed an error where from json returned a row that didn't have an id and could cause problems
+- fixed double init request in cases where `withRemult` accidentally wraps remult server
 
 ## [0.27.12] 2024-09-05